Added CLI flag to dump syscalls
This pull-request adds a new flag `-syscalls` to dump the list of
syscalls we've implemented.  Since some of these syscalls are fake
I've added a marker for that in our structure and ensured it is
displayed.

To avoid gaps in our syscall list I implemented two new syscalls:

* L_WRITE
  * Write a character to the printer.  (Use "print.log" file instead).
* DRV_SETRO
  * This should set a drive as read-only, instead it does nothing.
* F_ATTRIB
  * This should set file attributes, instead it does nothing.
